Refactor type imports to use specific types for commands, attendance, and user models; create new type files for better organization
