+++ src/models/types/user.types.ts
@@ -1,0 +1,5 @@
+export interface UserAttributes {
+    id: number;
+    discordId: string;
+    username: string;
+}
